Extracted classes from Utility.cs.
