+++ Foxtrot/Foxtrot/HelperMethods.cs
@@ -15,47 +15,10 @@
 using System.Diagnostics; // needed for Debug.Assert (etc.)
 using System.Compiler;
 using System.Diagnostics.Contracts;
-using System.Collections.Generic;
 using System;
 
 namespace Microsoft.Contracts.Foxtrot
 {
-    /// <summary>
-    /// A wrapper just so preamble code can be recognized at the beginning of a method.
-    /// Preamble code is
-    /// a) closure class creation/initialization
-    /// b) base ctor calls (in instance initializers)
-    /// c) field initialization (in instance initializers)
-    /// </summary>
-    public sealed class PreambleBlock : Block
-    {
-        public PreambleBlock(StatementList statements) : base(statements)
-        {
-        }
-    }
-
-    /// <summary>
-    /// Used in MoveNext methods in place of the original contracts
-    /// </summary>
-    public sealed class AssumeBlock : Block
-    {
-        public AssumeBlock(StatementList statements) : base(statements)
-        {
-        }
-    }
-
-    /// <summary>
-    /// Contains initialization code that needs to go before contracts and thus cannot go into the preamble block
-    /// For now it contains the following:
-    /// a) anonymous cached method delegate initialization to null.
-    /// </summary>
-    public sealed class InitBlock : Block
-    {
-        public InitBlock(StatementList statements) : base(statements)
-        {
-        }
-    }
-
     public static class HelperMethods
     {
         private static Identifier MulticastDelegateId = Identifier.For("MulticastDelegate");
@@ -2811,522 +2774,6 @@ namespace Microsoft.Contracts.Foxtrot
             return literal.Value;
         }
 
-        internal class FindClosurePartsToDuplicate : InspectorIncludingClosures
-        {
-            /// <summary>
-            /// Type containing the method containing the closure
-            /// </summary>
-            private TypeNode containingType;
-
-            internal MemberList MembersToDuplicate = new MemberList();
-
-            /// <summary>
-            /// 
-            /// </summary>
-            /// <param name="containingType">Type containing the method containing the closures</param>
-            public FindClosurePartsToDuplicate(TypeNode containingType, Method sourceMethod)
-            {
-                this.CurrentMethod = Unspecialize(sourceMethod);
-                this.containingType = Unspecialize(containingType);
-            }
-
-            private void TryAddTypeToMembersToDuplicate(TypeNode type)
-            {
-                TypeNode template = type;
-                while (template.Template != null)
-                {
-                    template = template.Template;
-                }
-
-                if (!this.MembersToDuplicate.Contains(template))
-                {
-                    this.MembersToDuplicate.Add(template);
-                }
-            }
-
-            public override void VisitTypeReference(TypeNode type)
-            {
-                if (type == null) return;
-
-                if (HelperMethods.IsClosureType(this.containingType, type))
-                {
-                    TryAddTypeToMembersToDuplicate(type);
-                }
-                else
-                {
-                    Debug.Assert(!type.Name.Name.Contains("DisplayClass"));
-                    //Console.WriteLine("Non-closure part: {0}", type.FullName);
-                }
-                
-                base.VisitTypeReference(type);
-            }
-
-            public override void VisitConstruct(Construct cons)
-            {
-                if ((cons.Type.IsDelegateType()))
-                {
-                    UnaryExpression ue = cons.Operands[1] as UnaryExpression;
-                    if (ue != null)
-                    {
-                        MemberBinding mb = ue.Operand as MemberBinding;
-                        if (mb != null)
-                        {
-                            Method m = (Method) mb.BoundMember;
-                            if (IsClosureMethod(this.containingType, m))
-                            {
-                                if (HelperMethods.IsClosureType(this.containingType, m.DeclaringType))
-                                {
-                                    // Roslyn-based compiler changed non-capturing lambda caching.
-                                    // Instead of storing delegate in a field like CS$<>9_CachedAnonymousMethodDelegate1
-                                    // Roslyn-based compiler generates a closure class called <>c with a set of instance
-                                    // methods and a singleton instance.
-                                    // This change allows to duplicate this new closure type
-                                    if (IsRoslynBasedStaticClosure(m.DeclaringType))
-                                    {
-                                        TryAddTypeToMembersToDuplicate(m.DeclaringType);
-                                    }
-                                }
-                                else
-                                {
-                                    // Logic for pre-Roslyn compiler.
-
-                                    // then there is no closure class, m is just a method the compiler
-                                    // added to the class itself
-
-                                    // we record the instance here, although we will eventually make a copy of the
-                                    // template by making the method generic in its parent type type-parameters.
-                                    MembersToDuplicate.Add(m);
-                                }
-                            }
-                            else
-                            {
-                                //Console.WriteLine("Not atomic closure part: {0}", m.FullName);
-                                var declaringTypeName = m.DeclaringType.Name.Name;
-                                var name = m.Name.Name;
-                                string message =
-                                    string.Format(
-                                        "DeclaringName should contain 'DisplayClass', <>c or Name should not have '__'. \r\nDeclaringTypeName: {0}, Name: {1}",
-                                        declaringTypeName, name);
-
-                                Debug.Assert(
-                                    declaringTypeName.Contains("DisplayClass") || declaringTypeName.Contains("<>c") ||
-                                    !name.Contains("__"), message);
-                            }
-                        }
-                    }
-                }
-
-                base.VisitConstruct(cons);
-            }
-
-            public override void VisitAssignmentStatement(AssignmentStatement assignment)
-            {
-                if (assignment == null) return;
-
-                Construct cons = assignment.Source as Construct;
-                if (cons == null) goto JustVisit;
-                
-                if (!(cons.Type.IsDelegateType())) goto JustVisit;
-                
-                UnaryExpression ue = cons.Operands[1] as UnaryExpression;
-                if (ue == null) goto JustVisit;
-                
-                MemberBinding mb = ue.Operand as MemberBinding;
-                if (mb == null) goto JustVisit;
-                
-                Method m = mb.BoundMember as Method;
-
-                if (m.IsStatic)
-                {
-                    mb = assignment.Target as MemberBinding;
-                    if (mb == null) goto JustVisit;
-
-                    if (mb.TargetObject != null) goto JustVisit;
-                    
-                    // Record the static cache field used to hold the static closure
-                    MembersToDuplicate.Add(mb.BoundMember);
-
-                    goto End;
-                }
-
-            JustVisit:
-                if (assignment.Source == null) goto JustVisit2;
-                
-                if (assignment.Source.NodeType != NodeType.Pop) goto JustVisit2;
-                
-                mb = assignment.Target as MemberBinding;
-                
-                if (mb == null) goto JustVisit2;
-                
-                if (mb.TargetObject != null) goto JustVisit2;
-                
-                if (mb.BoundMember == null) goto JustVisit2;
-                
-                if (Unspecialize(mb.BoundMember.DeclaringType) != this.containingType) goto JustVisit2;
-                
-                MembersToDuplicate.Add(mb.BoundMember);
-
-            JustVisit2:
-                ;
-
-            End:
-                base.VisitAssignmentStatement(assignment);
-            }
-        }
-
-        internal class DuplicatorForContractsAndClosures : Duplicator
-        {
-            protected Method sourceMethod;
-            protected Method targetMethod;
-            private Parameter replaceThisWithParameter;
-
-            /// <summary>
-            /// Nonnull if the contract is copied from a contract class
-            /// </summary>
-            private TypeNode contractClass;
-
-            private TypeNode contractClassToForward;
-            private TypeNode targetTypeToForwardTo;
-
-            public DuplicatorForContractsAndClosures(Module module, Method sourceMethod, Method targetMethod,
-                ContractNodes contractNodes)
-                : this(module, sourceMethod, targetMethod, contractNodes, true)
-            {
-            }
-
-            public DuplicatorForContractsAndClosures(Module module, Method sourceMethod, Method targetMethod,
-                ContractNodes contractNodes, bool mapParameters)
-                : base(module, targetMethod.DeclaringType)
-            {
-                this.sourceMethod = sourceMethod;
-                this.targetMethod = targetMethod;
-
-                this.RemoveNameForLocals = true;
-
-                Duplicator dup = this;
-
-                if (mapParameters)
-                {
-                    if (sourceMethod.ThisParameter != null)
-                    {
-                        if (targetMethod.ThisParameter != null)
-                        {
-                            dup.DuplicateFor[sourceMethod.ThisParameter.UniqueKey] = targetMethod.ThisParameter;
-                        }
-                        else
-                        {
-                            // target is a static wrapper. But duplicator cannot handle This -> Parameter conversion
-                            // so we handle it explicitly here in this visitor.
-                            replaceThisWithParameter = targetMethod.Parameters[0];
-                        }
-                    }
-
-                    if (sourceMethod.Parameters != null && targetMethod.Parameters != null
-                        && sourceMethod.Parameters.Count == targetMethod.Parameters.Count)
-                    {
-                        for (int i = 0, n = sourceMethod.Parameters.Count; i < n; i++)
-                        {
-                            dup.DuplicateFor[sourceMethod.Parameters[i].UniqueKey] = targetMethod.Parameters[i];
-                        }
-                    }
-                }
-
-                var originalType = IsContractTypeForSomeOtherType(sourceMethod.DeclaringType, contractNodes);
-                if (originalType != null)
-                {
-                    var contractType = this.contractClass = sourceMethod.DeclaringType;
-                    while (contractType.Template != null)
-                    {
-                        contractType = contractType.Template;
-                    }
-                    
-                    while (originalType.Template != null)
-                    {
-                        originalType = originalType.Template;
-                    }
-
-                    // forward ContractType<A,B> -> originalType<A',B'>
-                    this.contractClassToForward = contractType;
-                    this.targetTypeToForwardTo = originalType;
-
-                    //dup.DuplicateFor[contractType.UniqueKey] = originalType;
-                }
-            }
-
-            /// <summary>
-            /// If true, all names for duplicated local variables would be removed.
-            /// </summary>
-            public bool RemoveNameForLocals { get; set; }
-
-            public TypeNode PossiblyRemapContractClassToInterface(TypeNode candidate)
-            {
-                Contract.Requires(candidate != null);
-
-                var type = Unspecialize(candidate);
-                if (type == this.contractClassToForward)
-                {
-                    if (candidate.TemplateArguments != null)
-                    {
-                        var inst = this.targetTypeToForwardTo.GetTemplateInstance(candidate, candidate.TemplateArguments);
-                        return inst;
-                    }
-                    
-                    return this.targetTypeToForwardTo;
-                }
-
-                return candidate;
-            }
-
-            public override Expression VisitLocal(Local local)
-            {
-                var result = base.VisitLocal(local);
-                
-                var asLoc = result as Local;
-                if (asLoc != null && RemoveNameForLocals)
-                {
-                    asLoc.Name = null; // Don't clash with original local name
-                }
-                
-                return result;
-            }
-
-            public override Field VisitField(Field field)
-            {
-                field = base.VisitField(field);
-                
-                if (field == null) return field;
-
-                field.Type = PossiblyRemapContractClassToInterface(field.Type);
-                
-                return field;
-            }
-
-            public override TypeNode VisitTypeReference(TypeNode type)
-            {
-                if (type == null) return null;
-
-                var result = (TypeNode) this.DuplicateFor[type.UniqueKey];
-                
-                if (result != null) return result;
-                
-                result = base.VisitTypeReference(type);
-                
-                this.DuplicateFor[result.UniqueKey] = result;
-                
-                return result;
-            }
-
-            public override Member VisitMemberReference(Member member)
-            {
-                Method m = member as Method;
-                if (m != null && m.DeclaringType != null && m.DeclaringType == this.contractClass)
-                {
-                    // Find the corresponding reference in the interface/abstract class
-                    var implIntf = m.ImplementedInterfaceMethods;
-                    if (implIntf != null)
-                    {
-                        foreach (var im in implIntf)
-                        {
-                            return im;
-                        }
-                    }
-
-                    var impimplIntf = m.ImplicitlyImplementedInterfaceMethods;
-                    if (impimplIntf != null)
-                    {
-                        foreach (var im in impimplIntf)
-                        {
-                            return im;
-                        }
-                    }
-                }
-
-                return base.VisitMemberReference(member);
-            }
-
-            public override Method VisitMethod(Method method)
-            {
-                if (method == null) return null;
-
-                method = base.VisitMethod(method);
-
-                method.Template = null;
-                return method;
-            }
-
-            public override RequiresPlain VisitRequiresPlain(RequiresPlain plain)
-            {
-                if (plain == null) return null;
-
-                var result = base.VisitRequiresPlain(plain);
-                
-                // resanitize
-                result.UserMessage = ExtractorVisitor.FilterUserMessage(this.targetMethod, result.UserMessage);
-                
-                return result;
-            }
-
-            public override EnsuresNormal VisitEnsuresNormal(EnsuresNormal normal)
-            {
-                if (normal == null) return null;
-
-                var result = base.VisitEnsuresNormal(normal);
-                
-                result.UserMessage = ExtractorVisitor.FilterUserMessage(this.targetMethod, result.UserMessage);
-                
-                return result;
-            }
-
-            public override EnsuresExceptional VisitEnsuresExceptional(EnsuresExceptional exceptional)
-            {
-                if (exceptional == null) return null;
-
-                var result = base.VisitEnsuresExceptional(exceptional);
-                result.UserMessage = ExtractorVisitor.FilterUserMessage(this.targetMethod, result.UserMessage);
-                
-                return result;
-            }
-
-            public override Class VisitClass(Class Class)
-            {
-                Class = base.VisitClass(Class);
-                
-                Class.Template = null;
-
-                return Class;
-            }
-
-            public override Struct VisitStruct(Struct Struct)
-            {
-                Struct = base.VisitStruct(Struct);
-
-                Struct.Template = null;
-                
-                return Struct;
-            }
-
-            public override Expression VisitThis(This This)
-            {
-                var original = This;
-
-                var result = base.VisitThis(This);
-                
-                This = result as This;
-                
-                if (this.replaceThisWithParameter != null && This != null)
-                {
-                    if (original == sourceMethod.ThisParameter)
-                    {
-                        return this.replaceThisWithParameter;
-                    }
-                }
-
-                return result;
-            }
-
-            public override Expression VisitMethodCall(MethodCall call)
-            {
-                var result = base.VisitMethodCall(call);
-                
-                // for value type targets we may have to insert a constrained callvirt
-                call = result as MethodCall;
-                if (call == null) return result; // type changed
-                
-                MemberBinding memberBinding = call.Callee as MemberBinding;
-                if (memberBinding == null) return result;
-                
-                if (call.NodeType != NodeType.Callvirt)
-                {
-                    // check if we need to turn it into a call virt because we put in an abstract method call
-                    if (call.NodeType != NodeType.Call) return result;
-
-                    Method abstractMethod = memberBinding.BoundMember as Method;
-                    if (abstractMethod == null || !abstractMethod.IsAbstract) return result;
-                    
-                    call.NodeType = NodeType.Callvirt;
-                }
-
-                var target = memberBinding.TargetObject;
-                if (target == null) return result;
-
-                if (target is This && this.targetMethod.DeclaringType.IsValueType)
-                {
-                    call.Constraint = this.targetMethod.DeclaringType;
-                }
-                
-                if (target != targetMethod.ThisParameter && target != this.replaceThisWithParameter) return result;
-
-                TypeNode constraint = null;
-                if (targetMethod.DeclaringType.IsValueType)
-                {
-                    // it's a virtcall on "this" of a value type
-                    // add constraint
-                    constraint = targetMethod.DeclaringType;
-                }
-
-                if (constraint == null)
-                {
-                    constraint = IsConstrainedTypeVariable(this.replaceThisWithParameter);
-                }
-                
-                if (constraint != null)
-                {
-                    call.Constraint = constraint;
-                }
-                
-                return call;
-            }
-
-            private static TypeNode IsConstrainedTypeVariable(Parameter target)
-            {
-                if (target == null) return null;
-
-                Reference r = target.Type as Reference;
-                
-                if (r == null) return null;
-                
-                if (!(r.ElementType is ITypeParameter)) return null;
-                
-                return r.ElementType;
-            }
-
-            /// <summary>
-            /// Have to special case assignment of "this" param into other local or closure this-field when we have call-site wrappers
-            /// for constrained virtcalls. In this case, we have to insert a box.
-            /// </summary>
-            public override Statement VisitAssignmentStatement(AssignmentStatement assignment)
-            {
-                if (assignment == null) return null;
-
-                var result = base.VisitAssignmentStatement(assignment);
-                
-                assignment = result as AssignmentStatement;
-                
-                if (assignment != null && assignment.Target != null && assignment.Target.Type is Interface)
-                {
-                    if (assignment.Source != null & assignment.Source.Type is Reference)
-                    {
-                        var refType = (Reference) assignment.Source.Type;
-                        if (refType.ElementType is ITypeParameter)
-                        {
-                            // found a type mismatch
-                            assignment.Source =
-                                new BinaryExpression(
-                                    new AddressDereference(assignment.Source, refType.ElementType),
-                                    new Literal(refType.ElementType), NodeType.Box);
-                        }
-                    }
-                }
-
-                return result;
-            }
-
-            public virtual void SafeAddMember(TypeNode targetType, Member duplicatedMember, Member originalMember)
-            {
-                HelperMethods.SafeAddMember(targetType, duplicatedMember, originalMember);
-            }
-        }
-
         private static void FindClosureInitialization(Method containing, Block b, out Local closureLocal)
         {
             Contract.Requires(containing != null);
@@ -4061,7 +3508,7 @@ namespace Microsoft.Contracts.Foxtrot
         /// <summary>
         /// Renames duplicatedMember name if clashing (and it is a type)
         /// </summary>
-        private static void SafeAddMember(TypeNode targetType, Member duplicatedMember, Member original)
+        internal static void SafeAddMember(TypeNode targetType, Member duplicatedMember, Member original)
         {
             Debug.Assert(targetType.Template == null);
             
@@ -4747,186 +4194,6 @@ namespace Microsoft.Contracts.Foxtrot
             }
 
             return null;
-        }
-    }
-
-    /// <summary>
-    /// Can be used to find any member uses tagged with particular attributes (e.g. Model, RuntimeIgnored).
-    /// Goes into anonymous delegates as well.
-    /// </summary>
-    public class CodeInspector : Inspector
-    {
-        public static bool IsRuntimeIgnored(Node node, ContractNodes contractNodes, TypeNode referencingType, bool skipQuantifiers)
-        {
-            CodeInspector ifrv = new CodeInspector(
-                ContractNodes.RuntimeIgnoredAttributeName, contractNodes, referencingType, skipQuantifiers);
-            
-            ifrv.Visit(node);
-
-            return ifrv.foundAttribute;
-        }
-
-        public static bool UsesModel(Node node, ContractNodes contractNodes)
-        {
-            CodeInspector ifrv = new CodeInspector(ContractNodes.ModelAttributeName, contractNodes, null, false);
-            
-            ifrv.Visit(node);
-            
-            return ifrv.foundAttribute;
-        }
-
-        /// <summary>
-        /// True iff node should be ignored at runtime
-        /// </summary>
-        public bool foundAttribute;
-
-        private readonly Identifier attributeToFind;
-        private readonly ContractNodes contractNodes;
-        private readonly bool skipQuantifiers;
-        private readonly Stack<TypeNode> referencingType;
-        private bool generatedMethodContainsInvisibleMemberReference;
-
-        private CodeInspector(Identifier attributeToFind, ContractNodes contractNodes, TypeNode referencingType, bool skipQuantifiers)
-        {
-            this.foundAttribute = false;
-            
-            this.attributeToFind = attributeToFind;
-            this.contractNodes = contractNodes;
-            this.skipQuantifiers = skipQuantifiers;
-            this.referencingType = new Stack<TypeNode>();
-
-            if (referencingType != null)
-            {
-                this.referencingType.Push(referencingType);
-            }
-        }
-
-        public override void VisitMemberBinding(MemberBinding memberBinding)
-        {
-            if (memberBinding.BoundMember != null)
-            {
-                if (HelperMethods.HasAttribute(memberBinding.BoundMember.Attributes, this.attributeToFind))
-                {
-                    this.foundAttribute = true;
-                    return;
-                }
-
-                if (referencingType.Count > 0 &&
-                    !HelperMethods.IsVisibleFrom(memberBinding.BoundMember, this.referencingType.Peek()))
-                {
-                    this.generatedMethodContainsInvisibleMemberReference = true;
-                    this.foundAttribute = true;
-                    return;
-                }
-
-                Method referencedMethod = HelperMethods.Unspecialize(memberBinding.BoundMember as Method);
-                if (referencedMethod != null)
-                {
-                    if (HasAttribute(referencedMethod))
-                    {
-                        this.foundAttribute = true;
-                        return;
-                    }
-
-                    if (this.skipQuantifiers &&
-                        (this.contractNodes.IsForallMethod(referencedMethod) ||
-                         this.contractNodes.IsGenericForallMethod(referencedMethod) ||
-                         this.contractNodes.IsExistsMethod(referencedMethod) ||
-                         this.contractNodes.IsGenericExistsMethod(referencedMethod))
-                        )
-                    {
-                        this.foundAttribute = true;
-                        return;
-                    }
-
-                    // check if we deal with a contract method on an interface/abstract method that is annotated
-                    var origType = HelperMethods.IsContractTypeForSomeOtherTypeUnspecialized(referencedMethod.DeclaringType, this.contractNodes);
-                    
-                    if (origType != null)
-                    {
-                        var origMethod = HelperMethods.FindImplementedMethodSpecialized(origType, referencedMethod);
-                        if (HasAttribute(origMethod))
-                        {
-                            this.foundAttribute = true;
-                            return;
-                        }
-                    }
-                }
-            }
-
-            base.VisitMemberBinding(memberBinding);
-        }
-
-        private bool HasAttribute(Method referencedMethod)
-        {
-            if (referencedMethod != null)
-            {
-                // look for attribute on property
-                if (referencedMethod.IsPropertyGetter && referencedMethod.DeclaringMember != null)
-                {
-                    if (HelperMethods.HasAttribute(referencedMethod.DeclaringMember.Attributes, this.attributeToFind))
-                    {
-                        return true;
-                    }
-                }
-            }
-
-            return false;
-        }
-
-        // Handle bodies of anonymous delegates
-        public override void VisitConstruct(Construct cons)
-        {
-            if (cons.Type.IsDelegateType())
-            {
-                UnaryExpression ue = cons.Operands[1] as UnaryExpression;
-                if (ue == null) goto JustVisit;
-
-                MemberBinding mb = ue.Operand as MemberBinding;
-                if (mb == null) goto JustVisit;
-                
-                Method m = mb.BoundMember as Method;
-                
-                if (HelperMethods.IsCompilerGenerated(m))
-                {
-                    bool savedInvisibleMemberRef = this.generatedMethodContainsInvisibleMemberReference;
-                    try
-                    {
-                        this.generatedMethodContainsInvisibleMemberReference = false;
-                        
-                        var unspecedM = HelperMethods.Unspecialize(m);
-                        var unspecedT = unspecedM.DeclaringType;
-                        
-                        Contract.Assert(unspecedT.Template == null);
-                        
-                        this.referencingType.Push(unspecedT);
-                        this.VisitBlock(unspecedM.Body);
-                        
-                        if (this.generatedMethodContainsInvisibleMemberReference)
-                        {
-                            // remove method (and all containing closure methods)
-                            savedInvisibleMemberRef = true;
-                            for (int i = 0; i < unspecedT.Members.Count; i++)
-                            {
-                                if (unspecedT.Members[i] == unspecedM)
-                                {
-                                    unspecedT.Members[i] = null;
-                                    break;
-                                }
-                            }
-                        }
-                    }
-                    finally
-                    {
-                        this.generatedMethodContainsInvisibleMemberReference = savedInvisibleMemberRef;
-                        this.referencingType.Pop();
-                    }
-                }
-            }
-
-        JustVisit:
-            
-            base.VisitConstruct(cons);
         }
     }
 }
